backend/requirements.txt
flask
flask-cors
requests
pyorbital
python-dotenv
